requirements.txt
certifi==2024.8.30
packaging==24.1
pygame==2.6.1
openai==1.12.0
python-dotenv==1.0.0
PyOpenGL==3.1.6
numpy>=1.24.1
sounddevice>=0.4.6
pyaudio>=0.2.11
pydub>=0.25.1
websockets>=11.0.3
